pyproject.toml
[tool.poetry]
name = "herre-server"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
balder = {version = "0.1.22", extras = ["subscription"]}
django-avatar = "^7.0.0"
django-registration-redux = "^2.11"
django-cors-headers = "^3.13.0"
omegaconf = "^2.2.3"
rich = "^12.6.0"
django-health-check = "^3.17.0"
django-probes = "^1.7.0"
django-crispy-forms = "^1.14.0"
psycopg2 = "^2.9.4"
oauthlib = "^3.2.1"
jwcrypto = "^1.4.2"
django-rest-framework = "^0.1.0"
gunicorn = "^20.1.0"
pyjwt = "^2.5.0"
channels-redis = "^4.0.0"
django-storages = "^1.13.1"
boto3 = "^1.26.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
